Add anchor to scene path, and display stream anchor
This allows to play videos which are not supported by JWplayer. Stash will display links to the local file and the stream in the "File Info" panel. This allows to open the stream or local file directly in the browser (if supported) or the app that is registered for the respective file type (e.g. VLC).