Skip Intro for video previews
Entire credit goes to Fonzie#5266 on discord, but he suggested the above code as a way to skip the beginning portion of video previews as that is often disclaimers or title cards for the scene. Since this blanketly applies the skip, it would impact scenes without these intros, but I think the skip is small enough that it should not make a difference for those scenes. 

Code is what Fonzie suggested, but it seems to make sense to me unless it has impact on other areas as well. I just moved the twoPercent variable out of the for loop since I didn't think it needed to be calculated each time.