Add "Merge from" and "Merge into" actions to tags
Add two actions to the tag operations.
The first "Merge from" will merge  the given "source" tags into the
current tag. I.e. usages of that tag will be updated to the current tag,
aliases of those tags will be added to the current tag and the given
tag(names) will be added as alias as well.
The second, "Merge into", works the other way around. This means usages of
the current tag will be updated to the given tag, aliases of the current
tag will be added to the given tag, wnd the current tags name will be
added as alias as well to the given tag.