Optimize studio filter on performers
Rewrite studios filter for performers so that it is a lot faster.
Standalone queries have been tested by WithoutPants as part of #1483 and
show that instead of taking 2,5, to 3,5 seconds on a generated database
it takes less than half a second (possibly even less than a tenth of a
second).