Introduce an event dispatcher
Introduce package event with a dispatcher subsystem. Add this to the
manager and to the plugin subsystem.

Whenever the plugin subsystem execute a PostHook, we dispatch an
Change event on the event dispatcher bus. This currently has no
effect, but allows us to register subsystems on the event bus for
further processing. In particular, search.

By design, we opt to hook the plugin system and pass to the event bus
for now. One, it makes it easier to remove again, and two, the context
handling inside the plugin subsystem doesn't want to live on the
other side of an event bus.

While here, write a test for the dispatcher code.