Introduce some early reindexing code
Reindexing of scenes at the moment, because that's what we have. The
core idea is fairly simple: batch-process a table, a 1000 entries
at a time, index them. Replace the data loader every 10 rounds
(10k entries) so it doesn't grow too big.

While reindexing is ongoing, the online changemap is still being built
in the background. If reindexing takes more than the timer ticker,
it will fire immediately after. If reindexing takes more than twice
the timer ticker, the ticker protects against this and only fires once.