Simplify full reindexing, full reindex studios
The strategy is to fold reindexing into a worklist which we
process through systematically. This reduces the full reindexer
into a single loop, which then collapses the code to a far simpler
code path, where the only variance is a switch on the document type.

Use this new strategy to handle studios as well for full reindexing.