Refactor mapped scrapers some more
Simplify parsing of post processing actions. Collect all found
entries, and test that there's only a single one at the end of
the function. This simplifies the code flow quite a lot.

Start coalescing code paths between xpath and json scrapers. The
mapping interface is now named a docQueryer, because it is a really
a query abstraction over a document: either (X)HTML, or JSON.

Make document wrapping structures abstract. Return a docQueryer instead,
so that's the only interface used by the mapped code. This allows
for hiding of the underlying representational structure.

Use the name fields over split. In the golang strings library, the
function named field-something operates much like what we want, except
it works on runes rather than a string separation.