Introduce Cobra for the CLI
Use Cobra to manage the CLI. There are a couple of errors w.r.t.
the override system which needs to be configured correctly. The
viper instance we want to manipulate is the override instance.

However, this requires some restructuring of the configuration init
as it is something which should be lifted upwards a bit IMO.