Only run hash file setting migration if required
This should finally put to bed a persistent, latent bug where the hash file naming and calculate MD5 settings would inexplicably be changed. This post-migration was the suspect every time. This should now mean that the post-migration will only be run if the system was migrated from a schema version before the oshash changes were introduced.