Remove image data from studio and partial
Object structs should try to only include data that is used both for creation and querying. ImageBytes was not populated or used when reading studios. Removed from partial for consistency.

This results in more code at the client end to set images, but this repeated pattern should probably be abstracted into a studio/service function, rather than included in the sqlite code.