Expanding the Play & O Date Displays
expanded the display of the play/o dates in the history panel to use data from the play counter and last_played date. it is a slight temporary measure, as i am not able to access the playdate odate tables due to lacking the necessary graphql skills.
also reverted changing the spaces in 3 random files i had not meant to push.