Fix hyphen symbol being used instead of the appropriate emdash
The "-" symbol is a hyphen, it serves only a singular purpose: as a hyphen forming hyphenated words, i.e. "mother-in-law".

The correct grammatical symbol in this context is the emdash "—". Lazy folk will often substitute a hyphen with a space on either side into contexts where an emdash should be used.