Small CSS changes
Small CSS change to allow the `source-selector` to be brought to the front of the controls to allow people to select which source they would like.

At first, I was going more complex. I created functions that would disable the `big-buttons` if a menu was open and I got it mostly working but I felt like it was overkill when we can modify like 3-4 lines of CSS and get the same result.

fixes https://github.com/stashapp/stash/issues/4066