Updating Reload Scrapers formatting
Per convo with people on Discord. I have updated the Reload Scrapers UI. It now adds a button if the filter box appears and then the button extends and takes up the whole space if the filter box does not exist.

The CSS is... janky at best. Not too great with it so opinions and potential fixes would be greatly appriciated.