Bottom Page-Count button causes scenes page to reset to top
I think was was happening is the browser was trying to do too much at once (Rendering the popup and focusing the input simultaneously). I believe it was triggering a reflow and setting the site back to default aka: back to the top.

I set the timeout to 0 which  moves the execution to the next loop event. It gives the browser time to do one and then the other, not both at the same time.

I added some comments to what I did as well (Trying to get better at that).

I moved `onKeyPress` to `onKeyDown` due to the former being depreciated. Can also move this to another PR if requested so it's doesn't get lost in the sauce.