Update Yarn installation link
The existing link goes to a page that says:
```
These instructions only cover Yarn versions prior to 2.0. Those versions entered maintenance mode in January 2020 and will eventually reach their end-of-life in terms of support. Please see the main website for the most up-to-date documentation: yarnpkg.com/getting-started/migration.
```

This PR updates the link to go to the current installation instructions.